Changes from Barak review. including seperating diff_scanner from diff
