+++ catalog/diff.go
@@ -11,8 +11,14 @@ const (
 )
 
 type Difference struct {
-	Entry                // Partially filled.
+	Entry                // Partially filled. Path is always set.
 	Type  DifferenceType `db:"diff_type"`
+}
+
+type diffResultRecord struct {
+	TargetEntryNotInDirectBranch bool // the entry is reflected via lineage, NOT in the branch itself
+	Difference
+	EntryCtid *string // CTID of the modified/added entry. Do not use outside of catalog diff-by-iterators. https://github.com/treeverse/lakeFS/issues/831
 }
 
 func (d Difference) String() string {
